LayoutTests/imported/w3c/web-platform-tests/html/semantics/interactive-elements/the-dialog-element/inert-focus-in-frames-expected.txt



FAIL Tests inert node focusing across frames and iframes. assert_equals: iframe-input expected Element node <input id="iframe-input" class="target" type="date"></input> but got null
